refactor: move TestProcessor and ArraySizeProcessor to processors
